Confirm at y/n prompts before removals

+++ clean.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 function yn {
-    read -p "$1? [y/n] " reply
+    read -p "$1 [y/n] " reply
     if [[ $reply =~ ^[Yy] ]]; then
         return 0
     else
@@ -14,15 +14,25 @@ if [[ $(id -u) -ne 0 ]]; then
     exit 1
 fi
 
-rm -rf /Library/Application\ Support/{Logic,GarageBand}
-rm -rf /Applications/GarageBand.app
+if yn "Remove musical files (such as GarageBand, Logic loops)?"; then
+    rm -rf /Library/Application\ Support/{Logic,GarageBand}
+    rm -rf /Applications/GarageBand.app
+fi
 
-rm -rf /Applications/Microsoft\ Silverlight
+if yn "Remove Microsoft Silverlight?"; then
+    rm -rf /Applications/Microsoft\ Silverlight
+fi
 
-rm -rf /Library/iTunes
+if yn "Remove iTunes files?"; then
+    rm -rf /Library/iTunes
+fi
 
-rm -rf /Library/Desktop\ Pictures
+if yn "Purge factory desktop pictures?"; then
+    rm -rf /Library/Desktop\ Pictures
+fi
 
-rm -rf /Users/*/Photos/{Photo\ Booth\ Library,Photos\ Library}
+if yn "Remove Photo libraries?"; then
+    rm -rf /Users/*/Photos/{Photo\ Booth\ Library,Photos\ Library}
+fi
 
 # TODO: Optionally remove McAfee
